+++ Tests/Integration/DSCResources/IisSite.config.ps1
@@ -25,12 +25,16 @@ configuration IisSite_config
         $SkipRuleType,
 
         [Parameter()]
+        [string[]]
+        $SkipRuleSeverity,
+
+        [Parameter()]
         [hashtable]
         $Exception,
 
         [Parameter()]
-        [hashtable]
-        $BackwardCompatibilityException,
+        [string[]]
+        $ResourceParameters,
 
         [Parameter()]
         [object]
@@ -51,46 +55,14 @@ configuration IisSite_config
 
     Node localhost
     {
-        & ([scriptblock]::Create("
-            IisSite SiteConfiguration
-            {
-                IisVersion  = '$TechnologyVersion'
-                StigVersion = '$StigVersion'
-                $(if ($null -ne $WebAppPool)
-                {
-                   "WebAppPool = @($( ($WebAppPool | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
-                })
-                $( "WebSiteName = @($( ($WebSiteName | ForEach-Object {"'$PSItem'"}) -join ',' ))`n" )
-                $(if ($OrgSettings -is [hashtable])
-                {
-                    "Orgsettings = @{`n$($OrgSettings.Keys |
-                        ForEach-Object {"'{0}' = {1}{2} = '{3}'{4}`n" -f
-                            $PSItem, '@{', $($OrgSettings[$PSItem].Keys), $($OrgSettings[$PSItem][$OrgSettings[$PSItem].Keys]), '}'})}"
-                }
-                elseif ($null -ne $OrgSettings)
-                {
-                    "Orgsettings = '$OrgSettings'"
-                })
-                $(if ($null -ne $Exception)
-                {
-                    "Exception = @{`n$($Exception.Keys |
-                        ForEach-Object {"'{0}' = {1}{2} = '{3}'{4}`n" -f
-                            $PSItem, '@{', $($Exception[$PSItem].Keys), $($Exception[$PSItem][$Exception[$PSItem].Keys]), '}'})}"
-                })
-                $(if ($null -ne $BackwardCompatibilityException)
-                {
-                    "Exception = @{`n$($BackwardCompatibilityException.Keys |
-                        ForEach-Object {"'{0}' = {1}`n" -f $PSItem, $BackwardCompatibilityException[$PSItem]})}"
-                })
-                $(if ($null -ne $SkipRule)
-                {
-                    "SkipRule = @($( ($SkipRule | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
-                }
-                if ($null -ne $SkipRuleType)
-                {
-                    "SkipRuleType = @($( ($SkipRuleType | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
-                })
-            }")
-        )
+        $psboundParams = $PSBoundParameters
+        $psboundParams.IisVersion = $psboundParams['TechnologyVersion']
+        $psboundParams.Remove('TechnologyRole')
+        $psboundParams.Remove('ConfigurationData')
+        $psboundParams.Remove('TechnologyVersion')
+
+        $resourceParamString = New-ResourceParameterString -ResourceParameters $ResourceParameters -PSBoundParams $psboundParams
+        $resourceScriptBlockString = New-ResourceString -ResourceParameterString $resourceParamString -ResourceName IisSite
+        & ([scriptblock]::Create($resourceScriptBlockString))
     }
 }